Parse value sent to java application to float

+++ assignment-02/arduino/lib/tasks/MonitorTask.cpp
@@ -13,10 +13,7 @@ MonitorTask::MonitorTask() :
 void MonitorTask::tick() {
     int value = distanceToWaterLevel(_sonar.getDistance());
     float distance = sonarPulseToMeter(value);
-    // float t = value / 1000.0 / 1000.0 / 2;
-    // float distance = t * SOUND_SPEED;
-    // distance *= 100;
 
-    MsgService.sendMsg(String((int)distance));
+    MsgService.sendMsg(String(distance));
 }
